Updating description for PasswordSignInAsync
The current description is misleading for the 4th parameter in the method. It might be read as "if the lockoutOnFailure parameter is set to true, the user will be locked out, even if this is not their last possible attempt (for example if this is their 2nd attempt to sign in and the maximum is 5 they will still be locked out if the sign-in fails)", even though this is not true. The parameter only determines whether the failed access acount should increase by 1 if the sign-in fails.

I believe the description I gave fits better.